requirements.txt
defusedxml==0.4.1
Django==1.7.7
django-allauth==0.19.0
django-bootstrap3==5.0.3
django-configurations==0.8
django-crispy-forms==1.4.0
django-filter==0.9.2
django-polymorphic==0.6.1
djangorestframework==3.1.1
ipython==2.3.1
Markdown==2.6.1
oauthlib==0.7.2
Pillow==2.7.0
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
